+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@
 * You can configure a custom item sorting method from the Settings page.
 * Improved display of the account selection dropdown.
 * Fix pull from postmaster to clear exactly enough space, not too many, but also not too few.
+* In D2, "Make room for Postmaster" has been replaced with "Pull from Postmaster" which pulls all postmaster items we can onto your character.
 
 # 4.36.1